makefile base + nginx logs

+++ srcs/requirements/nginx/tools/entrypoint.sh
@@ -1,14 +1,36 @@
 #!/bin/sh
 set -e
 
-# Set server name in conf file
-echo "[nginx entrypoint] Setting up server name in nginx.conf..."
+# Check that global vars are set
 if [ -z "$SERVER_NAME" ]; then
-	echo "ERROR: SERVER_NAME not set"
+	echo "ERROR: SERVER_NAME not set. Exiting..."
 	exit 1
 fi
-sed -i "s/SERVER_NAME_PLACEHOLDER/${SERVER_NAME}/" /etc/nginx/nginx.conf.template
-mv /etc/nginx/nginx.conf.template /etc/nginx/nginx.conf
+
+# Set server name in conf file
+if [ -f /etc/nginx/nginx.conf.template ]; then
+	echo "[nginx entrypoint] Replacing vars in /etc/nginx/nginx.conf.template ..."
+	sed -i "s/SERVER_NAME_PLACEHOLDER/${SERVER_NAME}/" /etc/nginx/nginx.conf.template
+	echo "[nginx entrypoint] Creating /etc/nginx/nginx.conf ..."
+	mv /etc/nginx/nginx.conf.template /etc/nginx/nginx.conf
+else
+	echo "[nginx entrypoint] nginx.conf already created."
+fi
+
+# Generate SSL certificates
+if [ ! -d /etc/nginx/ssl ]; then
+	echo "[nginx entrypoint] Generating SSL certificates in /etc/nginx/ssl ..."
+	mkdir /etc/nginx/ssl
+	openssl req -x509 -nodes -days 365 -newkey rsa:2048 \
+		-keyout /etc/nginx/ssl/inception.key \
+		-out /etc/nginx/ssl/inception.crt \
+		-subj "/C=FR/ST=Catalunya/L=Barcelona/O=42/OU=42/CN=${SERVER_NAME}"
+	chmod 600 /etc/nginx/ssl/inception.key
+	chmod 644 /etc/nginx/ssl/inception.crt
+else
+	echo "[nginx entrypoint] SSL certificates already created."
+fi
 
 # Launch nginx in foreground (daemon off)
+echo "[nginx entrypoint] Launching nginx ..."
 exec nginx -g 'daemon off;'
